Organize functions in their own file
This really helps me find methods and make changes.
